+++ Engine/Player.h
@@ -6,8 +6,8 @@ class Player :
 public:
     bool inhibitUp = false, inhibitDown = false, inhibitRight = false, inhibitLeft = false, inhibitDrop = false;
     Player() = delete;
-    Player(int x, int y);
-    Player(int x, int y, int r, int g, int b);
+    Player(float x, float y);
+    Player(float x, float y, int r, int g, int b);
     void update();
 
     void draw(Graphics& gfx) const override;
